+++ pintos/include/threads/mmu.h
@@ -5,20 +5,20 @@
 #include <stdint.h>
 #include "threads/pte.h"
 
-typedef bool pte_for_each_func (uint64_t *pte, void *va, void *aux);
+typedef bool pte_for_each_func (uint8_t *pte, void *va, void *aux);
 
-uint64_t *pml4e_walk (uint64_t *pml4, const uint64_t va, int create);
-uint64_t *pml4_create (void);
-bool pml4_for_each (uint64_t *, pte_for_each_func *, void *);
-void pml4_destroy (uint64_t *pml4);
-void pml4_activate (uint64_t *pml4);
-void *pml4_get_page (uint64_t *pml4, const void *upage);
-bool pml4_set_page (uint64_t *pml4, void *upage, void *kpage, bool rw);
-void pml4_clear_page (uint64_t *pml4, void *upage);
-bool pml4_is_dirty (uint64_t *pml4, const void *upage);
-void pml4_set_dirty (uint64_t *pml4, const void *upage, bool dirty);
-bool pml4_is_accessed (uint64_t *pml4, const void *upage);
-void pml4_set_accessed (uint64_t *pml4, const void *upage, bool accessed);
+uint8_t *pml4e_walk (uint8_t *pml4, const uint8_t va, int create);
+uint8_t *pml4_create (void);
+bool pml4_for_each (uint8_t *, pte_for_each_func *, void *);
+void pml4_destroy (uint8_t *pml4);
+void pml4_activate (uint8_t *pml4);
+void *pml4_get_page (uint8_t *pml4, const void *upage);
+bool pml4_set_page (uint8_t *pml4, void *upage, void *kpage, bool rw);
+void pml4_clear_page (uint8_t *pml4, void *upage);
+bool pml4_is_dirty (uint8_t *pml4, const void *upage);
+void pml4_set_dirty (uint8_t *pml4, const void *upage, bool dirty);
+bool pml4_is_accessed (uint8_t *pml4, const void *upage);
+void pml4_set_accessed (uint8_t *pml4, const void *upage, bool accessed);
 
 #define is_writable(pte) (*(pte) & PTE_W)
 #define is_user_pte(pte) (*(pte) & PTE_U)
@@ -29,7 +29,7 @@ void pml4_set_accessed (uint64_t *pml4, const void *upage, bool accessed);
 /* Segment descriptors for x86-64. */
 struct desc_ptr {
 	uint16_t size;
-	uint64_t address;
+	uint8_t address;
 } __attribute__((packed));
 
 #endif /* thread/mm.h */
