+++ pintos/threads/thread.c
@@ -142,7 +142,7 @@ void thread_donate_chain(struct thread *donor) {
 // Global descriptor table for the thread_start.
 // Because the gdt will be setup after the thread_init, we should
 // setup temporal gdt first.
-static uint64_t gdt[3] = { 0, 0x00af9a000000ffff, 0x00cf92000000ffff };
+static uint8_t gdt[3] = { 0, 0x00af9a000000ffff, 0x00cf92000000ffff };
 
 /* Initializes the threading system by transforming the code
    that's currently running into a thread.  This can't work in
@@ -169,7 +169,7 @@ thread_init (void) {
 	 * The kernel will rebuild the gdt with user context, in gdt_init (). */
 	struct desc_ptr gdt_ds = {
 		.size = sizeof (gdt) - 1,
-		.address = (uint64_t) gdt
+		.address = (uint8_t) gdt
 	};
 	lgdt (&gdt_ds);
 
@@ -266,8 +266,8 @@ thread_create (const char *name, int priority,
 	/* Call the kernel_thread if it scheduled.
 	 * Note) rdi is 1st argument, and rsi is 2nd argument. */
 	t->tf.rip = (uintptr_t) kernel_thread;
-	t->tf.R.rdi = (uint64_t) function;
-	t->tf.R.rsi = (uint64_t) aux;
+	t->tf.R.rdi = (uint8_t) function;
+	t->tf.R.rsi = (uint8_t) aux;
 	t->tf.ds = SEL_KDSEG;
 	t->tf.es = SEL_KDSEG;
 	t->tf.ss = SEL_KDSEG;
@@ -525,7 +525,7 @@ static void init_thread (struct thread *t, const char *name, int priority) {
 	memset (t, 0, sizeof *t);
 	t->status = THREAD_BLOCKED;
 	strlcpy (t->name, name, sizeof t->name);
-	t->tf.rsp = (uint64_t) t + PGSIZE - sizeof (void *);
+	t->tf.rsp = (uint8_t) t + PGSIZE - sizeof (void *);
 	
 	t->priority = priority;                 // 유효 우선순위
 	t->base_priority = priority;            // 3️⃣ 원래 우선순위 저장(One)  => 복원 기준
@@ -573,7 +573,7 @@ do_iret (struct intr_frame *tf) {
 			"movw (%%rsp),%%es\n"
 			"addq $32, %%rsp\n"
 			"iretq"
-			: : "g" ((uint64_t) tf) : "memory");
+			: : "g" ((uint8_t) tf) : "memory");
 }
 
 /* Switching the thread by activating the new thread's page
@@ -588,8 +588,8 @@ do_iret (struct intr_frame *tf) {
    added at the end of the function. */
 static void
 thread_launch (struct thread *th) {
-	uint64_t tf_cur = (uint64_t) &running_thread ()->tf;
-	uint64_t tf = (uint64_t) &th->tf;
+	uint8_t tf_cur = (uint8_t) &running_thread ()->tf;
+	uint8_t tf = (uint8_t) &th->tf;
 	ASSERT (intr_get_level () == INTR_OFF);
 
 	/* The main switching logic.
